Code for Matrix traps
Given a 'N*M'matrix. Each cell in the matrix can have a value of either 0 or 1. You can start from any zero-valued cell and move to adjacent cells (up, down, left, or right) as long as the destination cell also has a value of 0.

An edge of the matrix is defined as the first row, the first column, the last row, or the last column. A trap cell is a cell where you can start but cannot reach an edge.

Return the number of trap cells.